# docs/swagger.yaml
definitions:
  common.MapAPIResp:
    properties:
      code:
        description: |-
          状态码
          Example: 200
        type: integer
      data:
        additionalProperties: {}
        description: |-
          数据
          可以是任意类型的数据
        type: object
      msg:
        description: |-
          信息
          Example: "success"
        type: string
    type: object
...
  common.Pag-sys_LoginRecordStandardOut:
    properties:
      items:
        description: 对象数组
        items:
          $ref: '#/definitions/sys.LoginRecordStandardOut'
        type: array
      page:
        description: |-
          当前页码
          Example: 1
        example: 1
        type: integer
      pages:
        description: |-
          总页数
          Example: 10
        example: 10
        type: integer
      size:
        description: |-
          每页数量
          Example: 10
        example: 10
        type: integer
      total:
        description: |-
          总记录数
          Example: 100
        example: 100
        type: integer
    type: object
...
  common.Pag-sys_UserDetailOut:
    properties:
      items:
        description: 对象数组
        items:
          $ref: '#/definitions/sys.UserDetailOut'
        type: array
      page:
        description: |-
          当前页码
          Example: 1
        example: 1
        type: integer
      pages:
        description: |-
          总页数
          Example: 10
        example: 10
        type: integer
      size:
        description: |-
          每页数量
          Example: 10
        example: 10
        type: integer
      total:
        description: |-
          总记录数
          Example: 100
        example: 100
        type: integer
    type: object
  errors.Error:
    properties:
      data:
        additionalProperties: {}
        type: object
      msg:
        type: string
      reason:
        $ref: '#/definitions/errors.ErrorReason'
    type: object
...
  sys.CreateUserDTO:
    properties:
      is_active:
        description: 是否激活
        type: boolean
      is_staff:
        description: 是否是工作人员
        type: boolean
      password:
        description: 密码
        maxLength: 20
        type: string
      role_id:
        description: 角色ID
        type: integer
      username:
        description: 用户名
        maxLength: 50
        type: string
    required:
    - password
    - role_id
    - username
    type: object
  sys.LoginDTO:
    properties:
      password:
        description: 密码
        maxLength: 20
        type: string
      username:
        description: 用户名
        maxLength: 50
        type: string
    required:
    - password
    - username
    type: object
  sys.LoginOut:
    properties:
      access_token:
        description: 登录令牌
        type: string
      refresh_token:
        description: 刷新令牌
        type: string
    type: object
  sys.LoginRecordStandardOut:
    properties:
      id:
        description: 唯一标识
        example: 1
        type: integer
      ip_address:
        description: IP地址
        example: 192.168.1.1
        type: string
      is_active:
        description: 登录状态
        example: true
        type: boolean
      login_at:
        description: 登录时间
        example: "2023-01-01 12:00:00"
        type: string
      user_agent:
        description: 用户浏览器信息
        example: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML,
          like Gecko) Chrome/96.0.4664.110 Safari/537.36
        type: string
      username:
        description: 名称
        example: judgement
        type: string
    type: object
  sys.LoginResp:
    properties:
      code:
        description: |-
          状态码
          Example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/sys.LoginOut'
        description: |-
          数据
          可以是任意类型的数据
      msg:
        description: |-
          信息
          Example: "success"
        type: string
    type: object
...
  sys.PagLoginRecordResp:
    properties:
      code:
        description: |-
          状态码
          Example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/common.Pag-sys_LoginRecordStandardOut'
        description: |-
          数据
          可以是任意类型的数据
      msg:
        description: |-
          信息
          Example: "success"
        type: string
    type: object
...
  sys.PagUserResp:
    properties:
      code:
        description: |-
          状态码
          Example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/common.Pag-sys_UserDetailOut'
        description: |-
          数据
          可以是任意类型的数据
      msg:
        description: |-
          信息
          Example: "success"
        type: string
    type: object
  sys.PatchPasswordDTO:
    properties:
      confirm_password:
        description: 确认密码
        type: string
      new_password:
        description: 新密码
        maxLength: 20
        type: string
      old_password:
        description: 原密码
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  sys.RefreshTokenDTO:
    properties:
      refresh_token:
        description: 刷新令牌
        type: string
    required:
    - refresh_token
    type: object
  sys.ResetPasswordDTO:
    properties:
      confirm_password:
        description: 确认密码
        type: string
      new_password:
        description: 新密码
        maxLength: 20
        type: string
    required:
    - confirm_password
    - new_password
    type: object
  sys.RoleBaseOut:
    properties:
      descr:
        description: 描述
        example: 用户管理
        type: string
      id:
        description: 唯一标识
        example: 1
        type: integer
      name:
        description: 名称
        example: 用户管理
        type: string
    type: object
...
  sys.UpdateUserDTO:
    properties:
      is_active:
        description: 是否激活
        type: boolean
      is_staff:
        description: 是否是工作人员
        type: boolean
      role_id:
        description: 角色ID
        type: integer
      username:
        description: 用户名
        maxLength: 50
        type: string
    required:
    - role_id
    - username
    type: object
  sys.UserDetailOut:
    properties:
      created_at:
        description: 创建时间
        example: "2023-01-01 12:00:00"
        type: string
      id:
        description: 唯一标识
        example: 1
        type: integer
      is_active:
        description: 是否激活
        example: true
        type: boolean
      is_staff:
        description: 是否是工作人员
        example: false
        type: boolean
      role:
        allOf:
        - $ref: '#/definitions/sys.RoleBaseOut'
        description: 角色
      updated_at:
        description: 更新时间
        example: "2023-01-01 12:00:00"
        type: string
      username:
        description: 用户名
        example: judgement
        type: string
    type: object
  sys.UserResp:
    properties:
      code:
        description: |-
          状态码
          Example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/sys.UserDetailOut'
        description: |-
          数据
          可以是任意类型的数据
      msg:
        description: |-
          信息
          Example: "success"
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/customer/me/password:
    patch:
      consumes:
      - application/json
      description: 本接口用于修改当前登录用户的密码
      parameters:
      - description: 修改用户密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys.PatchPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 密码修改成功
          schema:
            $ref: '#/definitions/common.MapAPIResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 修改当前用户密码
      tags:
      - 用户管理
  /api/v1/customer/me/record/login:
    get:
      consumes:
      - application/json
      description: 本接口用于查询当前登录用户的登录记录列表
      parameters:
      - description: |-
          登陆时间之后的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: after_login_at
        type: string
      - description: |-
          登陆时间之前的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: before_login_at
        type: string
      - description: 唯一标识
        in: query
        name: id
        type: integer
      - description: '"唯一标识列表(多个用,隔开)"'
        in: query
        maxLength: 100
        name: ids
        type: string
      - description: IP 地址
        in: query
        maxLength: 108
        name: ip_address
        type: string
      - description: 用户名
        in: query
        maxLength: 50
        name: name
        type: string
      - description: 分页页码
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 分页大小
        in: query
        minimum: 1
        name: size
        type: integer
      - description: 登陆状态
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户登录记录列表
          schema:
            $ref: '#/definitions/sys.PagLoginRecordResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 查询当前用户的登录记录列表
      tags:
      - 用户管理
...
  /api/v1/customer/user:
    get:
      consumes:
      - application/json
      description: 本接口用于查询用户列表
      parameters:
      - description: |-
          创建时间之后的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: after_created_at
        type: string
      - description: |-
          更新时间之后的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: after_updated_at
        type: string
      - description: |-
          创建时间之前的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: before_created_at
        type: string
      - description: |-
          更新时间之前的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: before_updated_at
        type: string
      - description: 唯一标识
        in: query
        name: id
        type: integer
      - description: '"唯一标识列表(多个用,隔开)"'
        in: query
        maxLength: 100
        name: ids
        type: string
      - description: 是否激活
        in: query
        name: is_active
        type: boolean
      - description: 是否是工作人员
        in: query
        name: is_staff
        type: boolean
      - description: 分页页码
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 角色ID
        in: query
        name: role_id
        type: integer
      - description: 分页大小
        in: query
        minimum: 1
        name: size
        type: integer
      - description: 用户名
        in: query
        maxLength: 50
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户列表
          schema:
            $ref: '#/definitions/sys.PagUserResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 查询用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 本接口用于新增用户
      parameters:
      - description: 创建用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 创建用户成功
          schema:
            $ref: '#/definitions/sys.UserResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 新增用户
      tags:
      - 用户管理
  /api/v1/customer/user/{id}:
    delete:
      consumes:
      - application/json
      description: 本接口用于删除指定ID的用户
      parameters:
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.MapAPIResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 本接口用于查询指定ID的用户
      parameters:
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户详情成功
          schema:
            $ref: '#/definitions/sys.UserResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 查询用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 本接口用于更新指定ID的用户
      parameters:
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      - description: 更新用户请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 更新用户成功
          schema:
            $ref: '#/definitions/sys.UserResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 用户管理
  /api/v1/customer/user/password/{id}:
    patch:
      consumes:
      - application/json
      description: 本接口用于重置指定ID的用户密码
      parameters:
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      - description: 重置用户密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys.ResetPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 密码重置成功
          schema:
            $ref: '#/definitions/common.MapAPIResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 重置用户密码
      tags:
      - 用户管理
  /api/v1/customer/user/record/login:
    get:
      consumes:
      - application/json
      description: 本接口用于查询用户登录记录列表
      parameters:
      - description: |-
          登陆时间之后的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: after_login_at
        type: string
      - description: |-
          登陆时间之前的记录 (RFC3339格式)
          example: 2023-01-01T00:00:00Z
        in: query
        name: before_login_at
        type: string
      - description: 唯一标识
        in: query
        name: id
        type: integer
      - description: '"唯一标识列表(多个用,隔开)"'
        in: query
        maxLength: 100
        name: ids
        type: string
      - description: IP 地址
        in: query
        maxLength: 108
        name: ip_address
        type: string
      - description: 用户名
        in: query
        maxLength: 50
        name: name
        type: string
      - description: 分页页码
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 分页大小
        in: query
        minimum: 1
        name: size
        type: integer
      - description: 登陆状态
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户登录记录列表
          schema:
            $ref: '#/definitions/sys.PagLoginRecordResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: 查询用户的登录记录列表
      tags:
      - 用户管理
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 本接口用于登陆
      parameters:
      - description: 登陆请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/sys.LoginResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      summary: 登陆接口
      tags:
      - 用户管理
...
  /api/v1/refresh/token:
    post:
      consumes:
      - application/json
      description: 本接口用于刷新令牌
      parameters:
      - description: 刷新令牌请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys.RefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新令牌成功
          schema:
            $ref: '#/definitions/sys.LoginResp'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.Error'
      summary: 刷新令牌接口
      tags:
      - 用户管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"